Remove the requirement to specify an API key + secret
Initially the plan was to only make it optional if a private key was
provided, but it's tough to do this without significant refactoring at
this time.

As the API responds nicely with a 401 and we handle errors in the lib,
removing this check seems OK to me. We can add the enforcing of
key/secret or private key at a later date

Resolves #159