Initial pass at media API implementation
There's a lot of room for improvement, but there's enough here to at
least try out the API.

* I've added `request` to handle the multipart form data for us. This
is the first step towards using `request` as our main HTTP client
* Media._makeRequest is required as Content-Type must be application/json
for all requests
* I've added a flag (`skipJsonDecoding`) to HTTPClient to allow us to
download a raw response (required for `media.download()`
out the mocks for `request`
* `media.download` implements both `If-Modified-Since` and
`If-None-Match` - https://ea.developer.nexmo.com/api/media#download-media-file
- but it is not documented. This is functionality we'll only expose if
people explicitly ask for it.
* Add support for mocking `request` when posting a file.  We check specifically
for the `postFile` method for now as that's the only thing that uses `request`.
This will need refactoring if we move away from HTTPClient in the future
* Add examples for Media API + supported list to README